+++ tests/unit/AuthorizationAdjustmentTest.php
@@ -21,10 +21,12 @@ class AuthorizationAdjustmentTest extends Setup
             'amount' => '100.00',
             'timestamp' => new \DateTime('2017-07-12 01:02:03'),
             'success' => true,
+            'processorResponseCode' => '1000',
+            'processorResponseText' => 'Approved',
         ];
 
         $authorizationAdjustment = Braintree\AuthorizationAdjustment::factory($authorizationAdjustmentParams);
 
-        $this->assertEquals('Braintree\AuthorizationAdjustment[amount=100.00, timestamp=Wednesday, 12-Jul-17 01:02:03 UTC, success=1]', (string) $authorizationAdjustment);
+        $this->assertEquals('Braintree\AuthorizationAdjustment[amount=100.00, timestamp=Wednesday, 12-Jul-17 01:02:03 UTC, success=1, processorResponseCode=1000, processorResponseText=Approved]', (string) $authorizationAdjustment);
     }
 }