+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## unreleased
+## Unreleased
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway
+* Add `processor_response_code` and `processor_response_text` to authorization adjustments subfield in transaction response.
 
 ## 3.34.0
 * Allow payee ID to be passed in options params for transaction create